control_server/requirements.txt
qiskit>=1.0
qiskit-aer>=0.14
fastapi>=0.110
uvicorn[standard]
httpx>=0.27
pydantic>=2.0
pytest>=8.0
pyyaml>=6.0
